revert a delete Drupal Bootstrap helper and rename DrupalAutoload
